% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_id_uf.R
\name{baixa_id_uf}
\alias{baixa_id_uf}
\title{Baixa UF}
\usage{
baixa_id_uf(dados_by_uf = TRUE)
}
\arguments{
\item{dados_by_uf}{é um argumento do tipo logical que se TRUE indica que a consulta será para os 27 estados, caso seja FALSE a consulta será por região Sul, Norte, Nordeste, Sudeste, Centro-Oeste e Brasil.}
}
\value{
tibble
}
\description{
Essa função baixa UF ou regiões com os respectivos códigos para baixar os dados em outra função.
}
